Integrate leiweibau/Pi.Alert-Satellite into Pi.Alert 18
RC0 Bugfix

chmod and chown have been added to the sudoer file to remove the error message "sudo: a terminal is required to read the password; either use the -S option to read from standard input or configure an askpass helper". These commands are required to reset the database to the correct permissions at the start of the scan if required.